Back/requirements.txt
asgiref==3.5.0
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
bidict==0.22.0
cffi==1.15.0
channels==3.0.4
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
Django==3.2.12
django-cors-headers==3.11.0
django-sslserver==0.22
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
dnspython==2.2.1
eventlet==0.33.0
greenlet==1.1.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
python-engineio==3.14.2
python-socketio==4.6.0
pytz==2021.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
zope.interface==5.4.0
